modified:   workspaces/DummyRosterWebApiImproved/DummyRoster.Common.EntityModel/Models/Customer.cs

+++ workspaces/DummyRosterWebApiImproved/DummyRoster.Common.EntityModel/Models/Customer.cs
@@ -11,7 +11,7 @@ namespace DummyRoster.Common.EntityModel.Models;
 public partial class Customer
 {
   [Key]
-  public long Id { get; set; }
+  public int Id { get; set; }
 
   [Column(TypeName = "varchar(32)")]
   public string Name { get; set; } = null!;
@@ -24,10 +24,10 @@ public partial class Customer
   public byte[]? Picture { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? Contact { get; set; }
+  public int? Contact { get; set; }
 
   [Column(TypeName = "INT")]
-  public long? Location { get; set; }
+  public int? Location { get; set; }
 
   [ForeignKey("Contact")]
   [InverseProperty("Customers")]
